requirements.txt
psycopg2-binary
jsonsimpleconfig
secureclientserverservice